Intern week 5 task
Learning Objectives:

Learn how to make programs do different things based on conditions.
Understand how to define and work with if statements.

Hands-On Coding Tasks:

Ask the user for a number. If the number is greater than 10, print "That's a big number!".
Ask the user for their age. If they are 18 or older, print "You are old enough to vote."
Ask for a fruit. If the fruit is "banana", print "Monkeys love bananas!"
Bonus Practice – Optional if you’d like to try

Check if a number is divisible by 3.
Check if a character is a vowel or consonant.
Check if a string starts with 'A'.
Check if age is eligible for voting.